requirements.txt
ultralytics>=8.2.0
opencv-python>=4.8.0
numpy>=1.24
tqdm>=4.66
pyyaml>=6.0
lxml>=5.2.0
shapely>=2.0.0
